Windows/E2E Test Case generator/requirements.txt
openai
langchain
langchain_openai
streamlit
python-dotenv
PyPDF2
torch
langsmith

-e .